[turboshaft] Prevent load elimination when loaded reps are not the same
https://crrev.com/c/4714607 fixed a bug where we were load eliminating
a Tagged load by replacing it with an untagged load. The fix added a
bitcast from untagged to tagged. However, this is somewhat unsafe in
general (because if a GC happened in between, we might need to
reload).

After looking more closely at the benchmark, the 2nd load was actually
unreachable: there were 2 map checks, the 1st one checking a string
map a loading the string length, and the 2nd one checking an array map
and loading the backing store. One of these 2 checks has to fail, and
the end of the code has to be unreachable.

This new CL prevents load eliminating loads with different
representations. All such cases should correspond to unreachable code
anyways. (we could replace the 2nd load by an `unreachable`, but it
might be safer to just not load-eliminate)

Bug: v8:12783, chromium:1467471
Change-Id: Iaf6b7d1126e3e5b021c931b44c6a35f480f71d47
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4720866
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89379}