[wasm][test] Allocate the far jump table big enough
cctests add functions after the NativeModule is already created. To
account for that the {NativeModule::ReserveCodeTableForTesting} method
allocate the near jump table big enough for the maximum number of
functions supported in tests (currently 10).

This CL extends this logic to do the same for the far jump table. Also,
when we later allocate more jump tables for additional code spaces we
make them just as big as the first jump tables. This fixes issues in
tests that overflow the first code space and then need far jumps between
code spaces.

We also adapt the initial code space estimate to take the maximum number
of functions into account instead of looking into the {WasmModule} which
will still be empty at this point.

R=manoskouk@chromium.org

Fixed: v8:14213
Change-Id: I24a7ca7d243d697c0fbcbeaffc9ca5f11b9b99b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4729567
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89406}