[wasm] JSToWasm wrapper is not subject to debugging
The JSToWasm was recently rewritten in Torque. Thereby the frame type
changed from JSToWasmFrame to TurbofanFrame. As a side effect, this
change made the JSToWasm wrapper subject to debugging. This side effect gets prevented by this CL by marking WasmExportedFunctions,
which contain the the JSToWasm wrapper, explicitly as not subject to
debugging.

R=clemensb@chromium.org

Bug: chromium:1470434
Change-Id: If47b908013004869a63d1bf1fbcceafc94a999a5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4756856
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89424}