[wasm] Don't cache the generic js-to-wasm wrapper
There is a cache for the js-to-wasm wrapper. So far any wrapper code got
cached there that was used, and therefore also the generic wrapper. This
is useless, as the generic wrapper is anyways always available.
With this CL the generic wrapper does not get cached anymore.

R=clemensb@chromium.org

Bug: v8:12926
Change-Id: I34a0a3ff071f33863a3093a9e87c3af569fe3698
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4757148
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89428}