[inspector] Always build preview for console API call arguments
console.table relies on the preview to build the actual table in
the frontend. This means it is currently necessary to have DevTools
open at the time the console message is logged.

This CL enables the preview for console messages sent to the frontend
on Runtime#enable. This allows users to call console.table with
DevTools closed.

Note that we are not sure why the preview was skipped in the first
place (and code archaeology didn't reveal it).

Also note that we use the console message's context to create the
temporary V8 JS objects required to serialize the preview.

Bug: chromium:1469536
Change-Id: I6af9f003508c4df6010df15037b89dfae727d9b3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4757149
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89431}