Fix Runtime_SimulateNewspaceFull for MinorMS
The old implementation was tailored for Scavenger and doesn't work for
MinorMS (specifically since it's assuming anything after top() is free
space).

Note that the new correct implementation is ~10x slower than the
previous one for MinorMS, but mostly because the previous one wasn't
really doing anything.
It's also 30% slower than with Scavenger, because (unlike Scavenger) it
needs to actually allocate new pages to fill space up to its target
capacity.
Since this is a testing only utility, the performance costs are deemed
acceptable.

Bug: chromium:1470515
Change-Id: Iee76b9ee1ae997973873daf3bcc7ef91dbb5ed33
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4769240
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89473}