[execution] Make Isolate::is_isolate_in_background_ atomic
Worker threads get the IsolateIn(Foreground/Background)Notification
from the main thread. This means the store on the main thread races
with loads of that flag on the worker thread.

While we could get away with relaxed as well, this flag is likely
not performance critical. So by using seqcst here, we are guaranteed
to always load the most recent store.

Bug: v8:14174
Change-Id: I84421d38e8c05588fffd1ac65c43271fdf15c47e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4773155
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89493}