[fuzzilli] Add more crash tests for fuzzing
In addition to the existing crash tests that verify that DCHECK and
CHECK failures are detected, this CL adds another one to verify that
invalid memory accesses are properly detected. This is useful for
example when fuzzing ASan builds to verify that ASans segv handler
properly reports crashes, and that these are detected by the fuzzer.

Bug: v8:10571
Change-Id: I48139fe4dfcb7493b3559b7104cba7c065799254
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4775583
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89519}