[wasm] Return early in case of an exception
After an exception has been thrown in the WebAssembly.Memory API
function we should return immediately such that the exception is
propagated correctly and we don't accidentally overwrite it with another
exception (or DCHECK in debug builds).

This CL also adds some more DCHECKs to document that we are returning
because an exception is pending at this point.

Plus some minor drive-by simplifications.

R=dlehmann@chromium.org

Bug: chromium:1472618
Change-Id: I252b6190c42aadae6790c51caabac3ded4201315
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4778659
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89531}