[wasm][cleanup] Provide a debug name for anonymous code
This is not printed anywhere, but helps to figure out what kind of code
object you are dealing with in a debugging session. For all "anonymous"
(non-wasm) code objects we only generated an "anonymous function" string
so far for the debug name. This CL changes that to look at the code kind
and generate a more precise string.

R=dlehmann@chromium.org

Bug: chromium:1407619
Change-Id: Ibae43b3e399266c896a6dd5d049268825dc25f45
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4790722
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89547}