[wasm][code-gc] Fix reference counting
If we do not add code to the {log_entry.code} vector, we also should not
increment its reference count. That reference count will never be
decremented again, so the code can never be garbage-collected.

This only happens if code is being logged, which is an exceptional case,
but we should fix this anyway.

Drive-by: Add some types (instead of auto) and some comments.

R=thibaudm@chromium.org

Change-Id: Id452e2e061ffad225e321a0dd7eee06dad6c6c5e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4790723
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89772}