[heap] Start/PauseObservablePause get timestamp as argument
Taking the timestamp as an argument helps to write tests for these
methods. The StartTracing/StopTracing helper functions for unit tests also start optionally taking a timestamp as an argument and pass it
to underlying GCTracer methods.

GCTracerTest::IncrementalMarkingSpeed uses it to manually pass in
timestamps. The unit test was broken with MemoryBalancer where a
DCHECK was violated.

Bug: v8:13842
Change-Id: I31050ab435b93d0fb48d61aa65fc4a664eae9e95
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4797628
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89570}