[profiler][etw] Enables filtering of ETW tracing by URL
Currently, when an ETW tracing session is started, it impacts all
renderer processes. In each renderer V8 emits ETW events to track all
the existing interpreted or jitted functions and for large websites the
emission of this initial burst of events can be relatively expensive.

This change allows to selectively enable ETW by URL, i.e., only on the
isolates that are running in a document that navigates to one of a set
of specified URLs.

Windows allows passing additional data to the ETW EnableCallback
https://learn.microsoft.com/en-us/windows/win32/api/evntprov/nc-evntprov-penablecallback

This data can be configured in a WPR (Windows Performance Recorder)
profile, adding a CustomFilter to an EventProvider like the following:

  <EventProvider Id="Edge_Stable" Name="57277741-3638-4A4B-BDBA-0AC6E45DA56C" Level="5">
    <CustomFilter Type="0x80000000" Value="AQABAAAAAAA..." />
  </EventProvider>

Where:
- Name="57277741-3638-4A4B-BDBA-0AC6E45DA56C" is the GUID of the V8
    ETW provider, (see src/libplatform/etw/etw-provider-win.h),
- Type="0x80000000" is EVENT_FILTER_TYPE_SCHEMATIZED,
- Value="AQABAAAAAA..." is a base64-encoded byte array that is
    base64-decoded by Windows and passed to the ETW enable callback in
    the 'PEVENT_FILTER_DESCRIPTOR FilterData' argument; see:
https://learn.microsoft.com/en-us/windows/win32/api/evntprov/ns-evntprov-event_filter_descriptor.

This array contains a struct EVENT_FILTER_HEADER followed by a
variable length payload, and as payload we pass string in JSON format,
with a list of regular expressions that should match the document URL
in order to enable ETW tracing:
  {
    "version": "1.0",
    "filtered_urls": [
        "https:\/\/.*\.chromium\.org\/.*", "https://v8.dev/", "..."
    ]
  }

The actual filtering will be implemented as a callback by the
embedder; see:
https://chromium-review.googlesource.com/c/chromium/src/+/4776268.

Bug: v8:11043
Change-Id: Icee297bec0560fd34b502d693d337cf69b707274
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4775987
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89616}