[handles] Refactor hierarchy of handles
This CL refactors the hierarchy of internal handles and implements a
few missing methods, to establish a better symmetry between (indirect)
Handle<T> and DirectHandle<T>. It introduces DirectHandleBase
(corresponding to HandleBase), which allows a few methods to be defined
in direct-handle.cc and resolves some existing compiler issues.

It also introduces IndirectHandle<T>, an alias for Handle<T>, to be
used for explicitly stating that a handle should be indirect during a
future migration from indirect to direct handles.

Bug: v8:13270
Change-Id: Ia7661aa397e5158e38d8c85f1bb410d874af4708
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4798491
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89573}