[heap] Use stack marker for conservative stack scanning
This CL refactors conservative stack scanning so that it always scans
from the stack marker to the stack start. It assumes that a stack marker
has already been set, when iterating through stack roots. The method
Stack::SetMarkerIfNeededAndCallback must be used to set the marker: it
only does so if the marker has not already been set. The callback is
required because the stack marker is set inside the trampoline for
saving callee-saved registers on the stack.

Using the stack marker instead of scanning the entire stack is useful
in cases where it is important to consistently scan the same part of
the stack successively. This CL addresses two such cases:

1. During GC, for marking and marking verification.
2. In the heap profiler, during snapshot generation, the GC and the
   filling of references should scan the same part of the stack.
   The same when querying for objects.

(A third case will be handled in a separate CL: during snapshot
serialization, GC and read-only promotion should scan the same part
of the stack or omit the stack entirely.)

The CL also cleans up the trampoline-related logic that was implemented
in the local heap, as well as the now redundant scan stack modes.

Bug: v8:13257
Bug: v8:14278
Change-Id: I36e68e4d5201def75ae9841dfcd9c0fd8fdf329e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4804204
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89610}