Substrings: when copying, check for one-byte range
When we decide to create a substring by copying (instead of creating
a SlicedString), and the source is a two-byte string, then with this
patch we check whether the to-be-copied range contains only one-byte
characters, and create a one-byte result string if so.
This is particularly important for Wasm, where our heuristics have a
much stronger bias for copying instead of creating slices, and hence
this can save a lot of memory. But while we're at it (and since some
of the implementation is shared anyway), we might as well support it
for JavaScript as well.

Bug: v8:12868, v8:14179
Change-Id: I09b518130b65167198b1cd34924353f61e4b1e85
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4807490
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89629}