[heap] Introduce Heap::MakeNewSpaceIterable method
Add method to make the new space iterable. This is useful for a
shared GC which needs the new space to be iterable.

So far we simply used Heap::MakeHeapIterable for convenience for this purpose. However, this method makes the whole heap iterable which is
not really necessary. But more importantly, Heap::MakeHeapIterable performs a lot of actions and e.g. creates handles in
GCTracer::StopCycle. Client isolates are parked during a shared GC and therefore shouldn't create handles.

Bug: chromium:1472372
Change-Id: Ia15ad5153e60e90c4732338a834a099e271caeec
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4813236
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89632}