[wasm] For I32PairToBigInt calls, lower parameter in graph construction
During graph construction, calls to the I32PairToBigInt builtin were set
up with a single parameter of type I64. This parameter was later lowered
to an I32 pair. To pass graph verification with the original graph, the
call was set up with the CallDescriptor of the I64ToBigInt builtin, and
this CallDescriptor also had to be changed in the lowering.

The lowering of the CallDescriptor is not trivial, as it requires to set
up a mapping from the original CallDescriptor to the new one.

With this CL, calls to the I32PairToBigInt builtin are set up with two
parameters of type I32. This is done by lowering the single I64
parameter already during graph construction and not only later in the
lowering. The lowering happens with TruncateI64ToI32 of a right shift by
32 bits, which both should end up as nops after the actual lowering.

R=jkummerow@chromium.org

Bug: v8:14108
Change-Id: I8e54cd7ab8dee6539d323a6f92a5767c033787d0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4813377
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89636}