[heap] Move stack marker when taking a heap snapshot
In https://crrev.com/c/4804204 a stack marker was introduced in
HeapSnapshotGenerator::GenerateSnapshot(), for ensuring that the GC
and the filling of references will scan the same part of the stack.
This CL moves the stack marker a bit higher in the call chain, to
HeapProfiler::TakeSnapshot, to guarantee that the part of the stack
above the marker will not be modified during snapshot generation.

Bug: v8:13257
Change-Id: I86d446ec709347032158fa929e2500ebfa083112
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4813324
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89641}