[heap] Fix MakeLinearAllocationAreaIterable for lab extensions
Usually when a LAB is allocated, we compute a limit for it and free the
remaining free space back to the free list.
PagedNewSpace supports LAB "extensions" by not freeing the memory and
instead keeping it allocated such that it could be used later without
going through the freelist again.

Making LAB iterable ignored that extra memory (between `limit()` and
`original_limit_relaxed()`) which meant that in some cases (e.g. when
`top() == limit()` when a client isolate tries to GC due to allocation
failure) calling `MakeLinearAllocationAreaIterable` would not create a
filler. If a shared GC happened during that time, the main isolate would
iterate the client isolate's new space and get lost because the LAB was
missing a filler.

Bug: chromium:1474285
Change-Id: I82c22c8ece6e515a5ef04a99bbfe81e6d2ea29b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4822566
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89682}