[cfi] Remove CodePageMemoryModificationScopeForPerf
We're down to two permission switches during IStream initialization and
reducing that to one doesn't make a difference in benchmarks so this
ForPerf scope is not needed anymore.
We need to put in a CodePageHeaderModificationScope in its place though
since the write barriers write to the headers. But this is Apple
silicon only and we're working on getting rid of the need for those
separately.

Change-Id: Ifb8473162ce1cb43a75beddfc4418471e6a3355b
Bug: v8:13355
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4790724
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#89708}