[heap] Do not allow heap expansion above max_old_generation_size()
This CL introduces a mutex that needs to be used when expanding
the old generation. This allows us to atomically check that the
OldGenerationCapacity() doesn't grow beyond max_old_generation_size().

However, this invariant is still violated between the GC safepoint
and the invocation of the near heap limit callbacks. This doesn't
allow us to add CHECKs for this.

Bug: v8:13842
Change-Id: I07050baebd29f0c0d3ca22b3d7211d2bfce91d83
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4822769
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89739}