[Assembler][ia32] Implement special lea instruction.
On ia32, there is no PIC mode.

The LoadLabelAddress macro generates a value that is the address of a
label in the code stream. It does so without actually embedding a label literal; which would result in non-relocatable code.

This is a fix of CL https://chromium-review.googlesource.com/c/v8/v8/+/4818822
which had to be reverted; and fixed.

The fix involved saving the ebx register since it is callee saved on
ia32.

Bug: v8:12191
Change-Id: Id4de98ceafb4cfe4eb0c4aa33112d91733e79761
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4834348
Reviewed-by: Adam Klein <adamk@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Francis McCabe <fgm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89756}