[wasm][revec] Fix control level calculation of LoopExit node
LoopExit node is directly connected to the Loop node. In a BFS search
this can make the control levels of the nodes after the loop smaller
than the control levels of nodes inside loop body. This CL fixes it
by ignoring this kind of edge.

Bug: v8:12716
Change-Id: I7920f52d409a017e3726349098c36b3be4f4361c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4837464
Reviewed-by: Tobias Tebbi <tebbi@chromium.org>
Commit-Queue: Hao A Xu <hao.a.xu@intel.com>
Cr-Commit-Position: refs/heads/main@{#89779}