[wasm] Log lazily compiled code in all isolates
We were missing to log lazily compiled code in other isolates that use
the same {NativeModule}. Instead we only checked the isolate in which
lazy compilation was triggered.

This CL fixes that by calling {WasmEngine::LogCode} from {CompileLazy},
which properly checks all isolates that use a {NativeModule}.
As that method is pretty expensive (it takes a Mutex to protect against
data races), we add a cheap-to-access atomic boolean to the
{NativeModule}. The {WasmEngine} updates this whenever a new isolate
uses a module, an isolate dies, or logging is enabled in an isolate.
This should mitigate any performance regressions.

R=thibaudm@chromium.org

Bug: chromium:1407619
Change-Id: Id21ed59ed14c6fa25c9d664c53989bb5aae09114
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4790962
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89778}